requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
coloredlogs==15.0.1
    # via onnxruntime
colorlog==6.9.0
    # via inferyx (pyproject.toml)
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fastapi==0.115.13
    # via inferyx (pyproject.toml)
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.58.4
    # via matplotlib
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via uvicorn
hf-xet==1.1.4
    # via huggingface-hub
huggingface-hub==0.33.0
    # via
    #   optimum
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via torch
kiwisolver==1.4.8
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via inferyx (pyproject.toml)
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.0
    # via
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   optimum
    #   pandas
    #   transformers
onnx==1.18.0
    # via inferyx (pyproject.toml)
onnxruntime==1.22.0
    # via inferyx (pyproject.toml)
optimum==1.26.1
    # via inferyx (pyproject.toml)
packaging==25.0
    # via
    #   huggingface-hub
    #   matplotlib
    #   onnxruntime
    #   optimum
    #   transformers
pandas==2.3.0
    # via inferyx (pyproject.toml)
pillow==11.2.1
    # via matplotlib
prometheus-client==0.22.1
    # via inferyx (pyproject.toml)
protobuf==6.31.1
    # via
    #   onnx
    #   onnxruntime
pydantic==2.11.7
    # via
    #   inferyx (pyproject.toml)
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via inferyx (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
rapidfuzz==3.13.0
    # via inferyx (pyproject.toml)
redis==6.2.0
    # via inferyx (pyproject.toml)
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   inferyx (pyproject.toml)
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
setuptools==80.9.0
    # via torch
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tokenizers==0.21.1
    # via transformers
torch==2.7.1
    # via optimum
tqdm==4.67.1
    # via
    #   inferyx (pyproject.toml)
    #   huggingface-hub
    #   transformers
transformers==4.52.4
    # via
    #   inferyx (pyproject.toml)
    #   optimum
typing-extensions==4.14.0
    # via
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   onnx
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.3
    # via inferyx (pyproject.toml)
